Fix: GUILE_EXTENSIONS_PATH to find library
Basically fixes the approach that we were previously trying with the minimum
viable example. Instead of trying to run everything from a single "run.sh"
script, and having to maintain a seperate guix.sh script for development
purposes, now "guix.sh" is to be invoked in order to set up a shell from which
we can launch the final program.